requirements.txt
Django==4.2.10
djangorestframework==3.14.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
django-cors-headers==4.3.0
dj-rest-auth==5.0.2
djangorestframework-simplejwt==5.3.1
google-api-python-client==2.116.0
google-auth-oauthlib==1.1.0
python-dateutil==2.8.2
django-filter==23.5
Pillow==10.1.0
drf-yasg==1.21.7
pytest==7.4.3
pytest-django==4.7.0 